Use pip-compile to enumerate requirements
``pip-compile`` (a command provided by ``pip-tools``) generates a full
frozen set of requirements and dependencies from a subset (usually, the
specific requirements the developer wanted to use directly and none of
their dependencies [which are really an implementation detail]). This
makes managing dependencies over time easier and provides a more stable
method of reproducing Python environments. For more information on this
tool, see the author's blog post [0] and mine [1].

[0] http://nvie.com/posts/better-package-management/
[1] https://www.jonafato.com/2015/12/15/Rethinking-requirements-txt