Spinner, fallback image, and noscript safe.
This uses a slightly different approach from before. This is not true 'lazy loading' but instead uses JS to add the spinner gif to the image background. If it loads successfully, the spinner is not visible. If it errors, then it's replaced by a default image and the spinner is removed.

This has the advantage of not breaking the expierence for users with JS disabled.