Upgrade pip-tools
The pinned version of `pip-tools` in `requirements/dev.in` is
incompatible with recent versions of `pip`. This change upgrades
`pip-tool` in isolation (as well as any dependencies of `pip-tools` that
are required by the dependency resolver) to make usage of `pip-compile`
work with newly-created (virtual) environments.

Additionally, it updates the dev requirements to include the fully
pinned requirements from the main `requirements.txt`. This ensures that
the dev environment stays in sync with the desired production
environment and that dev dependencies can be upgraded without
unintentionally upgrading non-dev dependencies.

Finally, it unpins `pip-tools` in the `.in` requirements file. In most
cases, pinning can be handled in the `.txt` file, and removing the pin
in this location allows developers to use the `--upgrade` and
`--upgrade-package` options with `pip-compile`.