Upgrade six
This is the first in a series of dependency upgrades to get the PyVideo
static site project up to date on modern releases. This may interact
with or eventually replace the work done in #303. The intent behind this
approach is to keep the diffs between upgrades to a minimum so that they
can easily be verified, explained, and linked to changelogs where
applicable.

This change upgrades `six` to its latest release. The upgrade produces
no difference in the rendered site output as verified by
[difftastic](https://difftastic.wilfred.me.uk/) with the command

```
difft main upgrade-six --skip-unchanged
```

on two full copies of the site created before and after installing the
new requirements file.

Other changes to the requirements file are limited to comments generated
by a newer version of `pip-compile`.