Upgrade blinker
This change is a followup to #341 and upgrades `blinker` to its
latest release. The upgrade produces no difference in the rendered site
output as verified by [difftastic](https://difftastic.wilfred.me.uk/)
with the command

```
difft main upgrade-blinker --skip-unchanged
```

on two full copies of the site created before and after installing the
new requirements file.