Upgrade markupsafe to 1.1.1
This is an incremental upgrade of `markupsafe`. This change produces
changes related to three videos, all of which are `alt` or `title`
attributes that were previously generated incorrectly based on video
titles and have been fixed by the change as noted in the
ampersand-related [changelog
entry](https://markupsafe.palletsprojects.com/en/latest/changes/#version-1-0)
for `markupsafe` 1.0.