Add simple redirect plugin
Add a redirect plugin to generate redirect pages based on a
"redirects.txt" file in the style of [Netlify] and [Cloudflare Pages]
redirects. This is motivated by a couple of needs:

1. Redirects / aliases for old event pages have been broken for some
   time (#61, #244, and others not documented in the issue tracker)
   despite having aliases defined in the data repository. The
   pelican-alias plugin appears to work for pages and articles
   (videos, in the case of PyVideo) but not categories (used here for
   events). (The pelican-alias plugin could be updated to support this
   use case, but it has not been updated in some time.)

2. Redirecting URLs to rendered pages that are not file-backed is
   useful, for example in the context of renaming and combining tags,
   redirecting speaker pages to updated URLs, or any other pages that
   are generated via strings contained within other page files.

The goals of this initial version are to fix the broken links mentioned
above and to unblock some efforts to clean up tags and improve speaker
name support. This has been tested via generating redirect links for the
currently-broken conference aliases, as shown in a followup commit in
the pull request associated with this change.

[Netlify]: https://docs.netlify.com/routing/redirects/
[Cloudflare Pages]: https://developers.cloudflare.com/pages/configuration/redirects/