Upgrade feedgenerator to 1.8
This upgrade produces different output for most (all?) feeds generated
for PyVideo. The changes made in this version are primarily related to
correcting the use of the `updated` vs. `published` elements, the use of
the `isPermalink` attribute, and the removal of Atom-specific elements
from RSS feeds.

These changes are noted in the [changelog], and the full list of changes
can be found in the [commit diff].

[changelog]: https://github.com/getpelican/feedgenerator/blob/c1e47f427d95ec71d5cfe651b268b063e5788741/CHANGELOG.md#180---2016-04-05
[commit diff]: https://github.com/getpelican/feedgenerator/compare/232fb7aafcde2c957f42d1d0adb0cbda13aeb495...b836c975f674f4c23dce2ef8600005e99de53ead